backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
insightface==0.7.3
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
ultralytics==8.0.145
torch==2.0.1
torchvision==0.15.2